+++ src/core/write.rs
@@ -11,7 +11,7 @@ impl<'a> QFilePath<'a> {
     /// # use qfile::QFilePath;
     /// # fn main() {
     /// // the real file path: `./FILE.txt`
-    /// let mut file = QFilePath::add_path("./file.txt");
+    /// let mut file = QFilePath::add_path("./file.txt").unwrap();
     /// file.auto_write("ok").unwrap();
     /// assert_eq(file.read().unwrap(),"ok");
     /// # }
@@ -20,38 +20,38 @@ impl<'a> QFilePath<'a> {
     ///
     /// - If the path exists, we work with the file (case insensitive)
     ///
+    /// |                            | Linux                        | Windows                      |
+    /// | -------------------------- | ---------------------------- | ---------------------------- |
+    /// | **The path we specified**: | `folder1/FolDER2/file.TXT`   | `folder1\FolDER2\file.TXT`   |
+    /// | **Real path** :            | `./Folder1/Folder2/file.txt` | `.\Folder1\Folder2\file.txt` |
+    /// | **Result** :               | `./Folder1/Folder2/file.txt` | `.\Folder1\Folder2\file.txt` |
     ///
-    /// | | |
-    /// |---|---|
-    /// | **The path we specified**: | `folder1/FolDER2/file.TXT` |
-    /// | **Real path** : | `./Folder1/Folder2/file.txt` |
-    /// | **Result** : | `./Folder1/Folder2/file.txt` |
     /// - If the file/path is not found, creates a new path with the file (*if initial path exists*)
     ///
-    /// | | |
-    /// |---|---|
-    /// | **The path we specified**: | `./folder/folder_new/file.txt` |
-    /// | **Real path** : | `./folder` |
-    /// | **Result** : | `./folder/folder_new/file.txt` |
+    /// |                            | Linux                          | Windows                        |
+    /// | -------------------------- | ------------------------------ | ------------------------------ |
+    /// | **The path we specified**: | `./folder/folder_new/file.txt` | `.\folder\folder_new\file.txt` |
+    /// | **Real path** :            | `./folder`                     | `.\folder`                     |
+    /// | **Result** :               | `./folder/folder_new/file.txt` | `.\folder\folder_new\file.txt` |
     ///
     /// - But if the initial path is different case of letters and a new file/folder is specified in the path, then a new path is created with the file
     ///
     /// ## Linux :
     ///
-    /// | | |
-    /// |---|---|
-    /// | **The path we specified**: | `./FOLDER/Folder_new/file.txt` |
-    /// |**Real path** : | `./folder` |
+    /// |                            |                                                         |
+    /// | -------------------------- | ------------------------------------------------------- |
+    /// | **The path we specified**: | `./FOLDER/Folder_new/file.txt`                          |
+    /// | **Real path** :            | `./folder`                                              |
     /// | **Result** :               | `./FOLDER/Folder_new/file.txt` - (**new created path**) |
-    /// |                            | `./folder` - (**original path**)
+    /// |                            | `./folder` - (**original path**)                        |
     ///
     /// ## Windows :
     ///
-    /// | | |
-    /// |---|---|
-    /// | **The path we specified**: | `./FOLDER/Folder_new/file.txt` |
-    /// |**Real path** : | `./folder` |
-    /// | **Result** :               | `./folder/Folder_new/file.txt` - (**real path**) |
+    /// |                            |                                                  |
+    /// | -------------------------- | ------------------------------------------------ |
+    /// | **The path we specified**: | `.\FOLDER\Folder_new\file.txt`                   |
+    /// | **Real path** :            | `.\folder`                                       |
+    /// | **Result** :               | `.\folder\Folder_new\file.txt` - (**real path**) |
     ///
     ///  ### Different behavior
     ///  > * The Windows file system treats file and directory names as **case insensitive**. `file.txt` and `FILE.txt` will be treated as equivalent files (Although the path is case insensitive in windows, you can return a case-sensitive path with : [`get_path_str()`](<struct.QFilePath.html#method.get_path_str>) or [`get_path_buf()`](<struct.QFilePath.html#method.get_path_buf>)).
@@ -138,7 +138,6 @@ impl<'a> QFilePath<'a> {
                 let path_without_file = fullpath.to_str().unwrap().rsplit_once(filename).unwrap().0;
                 {
                     self.user_path = PathBuf::from(path_without_file);
-                    // self.correct_path();
                     self.update_path = true;
                     self.file_name = PathBuf::from(filename);
                     self.flag = Flag::New;
@@ -149,10 +148,7 @@ impl<'a> QFilePath<'a> {
                     .unwrap();
                 Ok(())
             }
-            ErrorKind::PermissionDenied => {
-                panic!("Permission denied");
-            }
-            _ => panic!("other errors"),
+            _ => Err(err.into()),
         }
     }
     /// The same as [`auto_write()`](<struct.QFilePath.html#method.auto_write>), only the method for overwriting the data in the file
@@ -162,11 +158,28 @@ impl<'a> QFilePath<'a> {
     /// # fn main() {
     /// // the real file path: `file_Temp.txt`
     /// // file content: `1 2 3`
-    /// let mut file = QFilePath::add_path("File_temp.txt");
+    /// let mut file = QFilePath::add_path("File_temp.txt").unwrap();
     /// file.write_only_new("4 5 6").unwrap();
     /// assert_eq(file.read().unwrap(),"4 5 6");
     /// # }
     /// ```
+    /// ## Linux :
+    ///
+    /// |                            |                                                         |
+    /// | -------------------------- | ------------------------------------------------------- |
+    /// | **The path we specified**: | `./FOLDER/Folder_new/file.txt`                          |
+    /// | **Real path** :            | `./folder`                                              |
+    /// | **Result** :               | `./FOLDER/Folder_new/file.txt` - (**new created path**) |
+    /// |                            | `./folder` - (**original path**)                        |
+    ///
+    /// ## Windows :
+    ///
+    /// |                            |                                                  |
+    /// | -------------------------- | ------------------------------------------------ |
+    /// | **The path we specified**: | `.\FOLDER\Folder_new\file.txt`                   |
+    /// | **Real path** :            | `.\folder`                                       |
+    /// | **Result** :               | `.\folder\Folder_new\file.txt` - (**real path**) |
+    /// 
     pub fn write_only_new(&mut self, text: &str) -> Result<(), io::Error> {
         self.flag = Flag::New;
         if let Err(err) = self.auto_write(text) {