+++ src/dpds_path/mod.rs
@@ -2,6 +2,8 @@ pub use lazy_static::{__Deref, lazy_static};
 pub use regex::Regex;
 pub use std::{
     env,
+    error::Error,
+    fmt::Display,
     fs::{self, DirBuilder, File, OpenOptions},
     io::{self, ErrorKind, Read},
     path::{Path, PathBuf},