+++ tests/qfile-2_0.rs
@@ -33,7 +33,7 @@ fn pwmf(main_folder: &str, path: &str) -> String {
 fn unix_test_path_0_part1() {
     let main_folder = TestFolder::new(".Polygon").folder;
     let path = pwmf(&main_folder, "/file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     if let Ok(()) = unix_test_path_0_part2(&path) {
         assert_eq!(file.read().unwrap(), "okok");
@@ -44,7 +44,7 @@ fn unix_test_path_0_part1() {
     panic!(":: ERROR - part2 from first_slash_add1");
 }
 fn unix_test_path_0_part2(path: &str) -> Result<(), std::io::Error> {
-    QFilePath::add_path(path).auto_write("ok")
+    QFilePath::add_path(path).unwrap().auto_write("ok")
 }
 
 #[cfg(target_family = "unix")]
@@ -53,7 +53,7 @@ fn unix_test_path_0_part2(path: &str) -> Result<(), std::io::Error> {
 fn unix_test_path_1() {
     let main_folder = TestFolder::new("../delete me").folder;
     let path = pwmf(&main_folder, "/file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.auto_write("ok").unwrap();
     assert_eq!(file.read().unwrap(), "ok");
     delete_item(&main_folder);
@@ -65,6 +65,7 @@ fn unix_test_path_1() {
 // root folder
 fn unix_test_path_2() {
     QFilePath::add_path("/usr/invalid_file.txt")
+        .unwrap()
         .auto_write("delete this file ")
         .unwrap();
 }
@@ -75,7 +76,7 @@ fn unix_test_path_2() {
 fn unix_test_path_3() {
     let main_folder = TestFolder::new("Polygon").folder;
     let path = pwmf(&main_folder, "/a/b/c/file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     assert_eq!(file.read().unwrap(), "ok");
     delete_item(&main_folder);
@@ -87,10 +88,10 @@ fn unix_test_path_3() {
 fn unix_test_path_4() {
     let main_folder = TestFolder::new("Polygon").folder;
     let path = pwmf(&main_folder, "/a/B/c/file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("").unwrap();
     let find_path = format!("{}{}", main_folder.to_lowercase(), "/A/B/c/file.txt");
-    let mut find = QFilePath::add_path(&find_path);
+    let mut find = QFilePath::add_path(&find_path).unwrap();
     assert_eq!(find.get_path_str(), format!("./{}", path));
     delete_item(&main_folder);
 }
@@ -100,7 +101,7 @@ fn unix_test_path_4() {
 fn unix_test_path_5() {
     let main_folder = TestFolder::new("./Polygon").folder;
     let path = pwmf(&main_folder, "/a/b/c/file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     assert_eq!(file.read().unwrap(), "ok");
     delete_item(&main_folder);
@@ -111,7 +112,7 @@ fn unix_test_path_5() {
 fn unix_test_path_6() {
     let file_name = TestFolder::generate(15);
     let file_name = format!("{}.txt", file_name);
-    let mut file = QFilePath::add_path(&file_name);
+    let mut file = QFilePath::add_path(&file_name).unwrap();
     file.auto_write("").unwrap();
     if let Err(_) = fs::remove_file(&file_name) {
         panic!("unix_test_path_6 :\nError delete file")
@@ -123,7 +124,7 @@ fn unix_test_path_6() {
 fn windows_test_path_0_part1() {
     let main_folder = TestFolder::new(".Polygon").folder;
     let path = pwmf(&main_folder, "\\file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     if let Ok(()) = windows_test_path_0_part2(&path) {
         assert_eq!(file.read().unwrap(), "okok");
@@ -135,7 +136,7 @@ fn windows_test_path_0_part1() {
 }
 #[allow(dead_code)]
 fn windows_test_path_0_part2(path: &str) -> Result<(), std::io::Error> {
-    QFilePath::add_path(path).auto_write("ok")
+    QFilePath::add_path(path).unwrap().auto_write("ok")
 }
 
 #[cfg(target_family = "windows")]
@@ -144,7 +145,7 @@ fn windows_test_path_0_part2(path: &str) -> Result<(), std::io::Error> {
 fn windows_test_path_1() {
     let main_folder = TestFolder::new("..\\delete me").folder;
     let path = pwmf(&main_folder, "\\file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.auto_write("ok").unwrap();
     assert_eq!(file.read().unwrap(), "ok");
     delete_item(&main_folder);
@@ -156,6 +157,7 @@ fn windows_test_path_1() {
 // root folder
 fn windows_test_path_2() {
     QFilePath::add_path("C:\\Windows\\System32\\invalid_file.txt")
+        .unwrap()
         .auto_write("delete this file ")
         .unwrap();
 }
@@ -166,7 +168,7 @@ fn windows_test_path_2() {
 fn windows_test_path_3() {
     let main_folder = TestFolder::new("Polygon").folder;
     let path = pwmf(&main_folder, "\\a\\b\\cfile.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     assert_eq!(file.read().unwrap(), "ok");
     delete_item(&main_folder);
@@ -178,7 +180,7 @@ fn windows_test_path_3() {
 fn windows_test_path_4() {
     let main_folder = TestFolder::new("D:\\Polygon").folder;
     let path = pwmf(&main_folder, "\\a\\b\\cfile.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     assert_eq!(file.read().unwrap(), "ok");
     delete_item(&main_folder);
@@ -189,11 +191,11 @@ fn windows_test_path_4() {
 fn windows_test_path_5() {
     let main_folder = TestFolder::new(".Polygon").folder;
     let path = pwmf(&main_folder, "\\FOlder\\folder\\file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     // assert_eq!(file.read().unwrap(), "ok");
     let new_path = path.to_string();
-    let mut file = QFilePath::add_path(&new_path);
+    let mut file = QFilePath::add_path(&new_path).unwrap();
     assert_eq!(file.get_path_str(), format!(".\\{}", path));
     delete_item(&main_folder);
 }
@@ -203,7 +205,7 @@ fn windows_test_path_5() {
 fn windows_test_path_6() {
     let file_name = TestFolder::generate(15);
     let file_name = format!("{}.txt", file_name);
-    let mut file = QFilePath::add_path(&file_name);
+    let mut file = QFilePath::add_path(&file_name).unwrap();
     file.auto_write("").unwrap();
     if let Err(err) = fs::remove_file(&file_name) {
         panic!("windows_test_path_6 :\nError {{{}}}", err)
@@ -214,11 +216,11 @@ fn windows_test_path_6() {
 fn windows_test_path_7() {
     let main_folder = TestFolder::new("D:\\Polygon").folder;
     let path = pwmf(&main_folder, "\\FOlder\\folder\\file.txt");
-    let mut file = QFilePath::add_path(&path);
+    let mut file = QFilePath::add_path(&path).unwrap();
     file.write_only_new("ok").unwrap();
     // assert_eq!(file.read().unwrap(), "ok");
     let new_path = path.to_string();
-    let mut file = QFilePath::add_path(&new_path);
+    let mut file = QFilePath::add_path(&new_path).unwrap();
     assert_eq!(file.get_path_str(), path);
     delete_item(&main_folder);
 }